This is the solution to the first part in the problem we had been given a chunk of data containing ( and (. ( means we go up 1 floor and ) means we go down a floor.
in the solution we just make a global variable named floor and store the floor number in that variable.
we then use a for loop to iterate through the data and add 1 floor if python reads a ( and substract 1 floor if python reads ).
and then just print the floor santa is on at the end i.e. our part 1 answer.